+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## vNext
 
 - Rename --mirror to --mirror-patterns
+- Add --folder-symlink-as-destination
 - Log "Running dry" only when there is something not to execute
 - Publish on PyPI
 - Use python-poetry